[TSK]BROS-59-criar-view-de-gerenciamento-de-estoque [WIP]Resolvendo conflitos, e agora puxa os produtos para o estoque do banco
