+++ BROS-ECommerce.Domain/Entities/Estoque.cs
@@ -1,11 +1,6 @@
 ﻿using BROS_ECommerce.Domain.Interfaces;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BROS_ECommerce.Domain.Entities
 {
